requirements3.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
Authlib==1.5.1
av==14.3.0
babel==2.16.0
banks==2.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
executing==2.1.0
fastapi==0.115.12
fastjsonschema==2.20.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
gast==0.6.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
google-genai==1.5.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
livekit==1.0.6
livekit-agents==1.0.17
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.17
livekit-plugins-deepgram==1.0.17
livekit-plugins-google==1.0.11
livekit-plugins-groq==1.0.17
livekit-plugins-openai==1.0.17
livekit-plugins-silero==1.0.17
livekit-plugins-turn-detector==1.0.17
livekit-protocol==1.0.2
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index==0.12.34
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.34.post1
llama-index-embeddings-gemini==0.3.2
llama-index-embeddings-huggingface==0.5.3
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.21
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nnfs==0.5.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.5
onnxruntime==1.21.1
openai==1.76.0
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pypdf==5.4.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2024.2
pywin32==307
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.14.1
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-protobuf==4.25.0.20240417
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.20.0
zipfile36==0.1.3
